create newsItemList
