+++ src/components/NewsItemList.css
@@ -1,0 +1,3 @@
+.news-list-container > div {
+  margin-bottom: 16px;
+}
